Doubly Circular Linked List
The ticket booking system of Cinemax theater has to be implemented using C++ program. There are 10 rows and 7 seats in each row. Doubly circular linked list has to be maintained to keep track of free seats at rows. Assume some random booking to start with. Use array to store pointers (Head pointer) to each row.  On demand  
a) The list of available seats is to be displayed 
b) The seats are to be booked c)  The booking can be cancelled.